Move methods for creating CompositeBuffer to BufferAllocator
Motivation:

At the moment creation of CompositeBuffer is done via a static method. We should better add a method to the BufferAllocator for creation these buffers as this will allow users to easily decorate all type of buffers if needed.

Modifications:

- Move methods to BufferAllocator
- Adjust tests

Result:

More consistent API and possible to easily decorate